requirements.txt
Django==1.7
Markdown==2.5.2
Pillow==2.6.1
Werkzeug==0.9.6
argparse==1.2.2
beautifulsoup4==4.3.2
dj-static==0.0.6
django-classy-tags==0.5.1
django-extensions==1.4.6
django-grappelli==2.6.3
django-htmlmin==0.7.0
django-markitup==2.2.2
django-model-utils==2.2
django-widget-tweaks==1.3
djrill==1.2.0
easy-thumbnails==2.2
gunicorn==19.1.1
hiredis==0.1.5
html5lib==0.999
psycopg2==2.5.4
redis==2.10.3
requests==2.4.3
six==1.8.0
static3==0.5.1
virtualenv==1.11.6
wsgiref==0.1.2
